update issue 9
ChildWindows are added for the specified features. But this is only a cosmetic change.
For instance, if you click on the "add patient" button, it'll navigate to an empty
page with in the middle the ChildWindow.

I've created a new issue with low priority to build a generic system.

new issue
Summary: nDoctor must have a built in facility to open generic modal child windows.
Status: Accepted
Labels: Type-Enhancement Priority-Low Component-UI